Status Task implementation, partly

+++ main/common/esp/free_rtos/free_rtos_if.hpp
@@ -34,7 +34,7 @@ typedef struct tskTaskControlBlock * TaskHandle_t;
 #define pdFAIL  ( pdFALSE )
 
 // Queue types (no function stubs — use RtosQueueStub.cpp in test builds)
-typedef uint16_t TickType_t;
+typedef uint32_t TickType_t;
 #define CONFIG_FREERTOS_HZ 100
 #define configTICK_RATE_HZ CONFIG_FREERTOS_HZ
 #define pdMS_TO_TICKS( xTimeInMs ) \
